Move DataDetector into own file
